[jak3] implement pckernel
Most debug features do not work, but that's fine.
