+++ decompiler/config/jak3/all-types.gc
@@ -37208,7 +37208,7 @@
   :size-assert         #x1d
   :flag-assert         #x90000001d
   (:methods
-    (new (symbol type string int (function object object)) _type_) ;; 0
+    (new (symbol type string object (function object object)) _type_) ;; 0
     )
   )
 
@@ -37230,7 +37230,7 @@
   :size-assert         #x20
   :flag-assert         #x900000020
   (:methods
-    (new (symbol type string int (function object debug-menu-msg object)) _type_) ;; 0
+    (new (symbol type string object (function object debug-menu-msg object)) _type_) ;; 0
     )
   )
 
@@ -38950,7 +38950,6 @@
 ;; part-tester                    ;;
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 
-#|
 (deftype part-tester (process)
   ((root      trsqv                     :offset-assert 128) ;; guessed by decompiler
    (part      sparticle-launch-control  :offset-assert 132) ;; guessed by decompiler
@@ -38963,12 +38962,13 @@
     part-tester-idle ;; associated process guessed by decompiler, old: (state part-tester)
     )
   )
-|#
 
-;; (define-extern *part-tester-name* object) ;; string
-;; (define-extern part-tester-init-by-other function) ;; (function vector none :behavior process-drawable)
-;; (define-extern *debug-part-dead-pool* object) ;; dead-pool
-;; (define-extern start-part function) ;; (function none)
+(define-extern *part-tester-name* string)
+(define-extern part-tester-init-by-other (function vector none :behavior process-drawable))
+(define-extern *debug-part-dead-pool* dead-pool)
+(define-extern start-part (function none))
+(define-extern *part-tester* (pointer part-tester))
+(define-extern part-tester-idle (state part-tester))
 
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;; manipulator                    ;;
