+++ decompiler/config/jak3/ntsc_v1/label_types.jsonc
@@ -957,5 +957,9 @@
     ["L13", "vector"],
     ["L12", "vector"],
     ["L11", "vector"]
+  ],
+  "part-tester": [
+    ["L49", "uint64", true],
+    ["L44", "uint64", true]
   ]
 }
