+++ decompiler/config/jak3/ntsc_v1/type_casts.jsonc
@@ -6619,5 +6619,6 @@
   "(method 62 v-fox)": [[[4, 48], "s5", "collide-shape-prim-group"]],
   "(method 62 v-rhino)": [[[4, 41], "s5", "collide-shape-prim-group"]],
   "(method 62 v-mirage)": [[[4, 48], "s5", "collide-shape-prim-group"]],
-  "(method 62 v-x-ride)": [[[4, 48], "s5", "collide-shape-prim-group"]]
+  "(method 62 v-x-ride)": [[[4, 48], "s5", "collide-shape-prim-group"]],
+  "(code part-tester-idle)": [[[6, 22], "s5", "process-drawable"]]
 }
