+++ game/CMakeLists.txt
@@ -128,6 +128,7 @@ set(RUNTIME_SOURCE
         kernel/jak3/klink.cpp
         kernel/jak3/klisten.cpp
         kernel/jak3/kmachine.cpp
+        kernel/jak3/kmachine_extras.cpp
         kernel/jak3/kmalloc.cpp
         kernel/jak3/kmemcard.cpp
         kernel/jak3/kprint.cpp
