+++ game/kernel/jak3/kmachine.cpp
@@ -21,6 +21,7 @@
 #include "game/kernel/jak3/kboot.h"
 #include "game/kernel/jak3/kdgo.h"
 #include "game/kernel/jak3/klisten.h"
+#include "game/kernel/jak3/kmachine_extras.h"
 #include "game/kernel/jak3/kmalloc.h"
 #include "game/kernel/jak3/kscheme.h"
 #include "game/kernel/jak3/ksound.h"
@@ -351,38 +352,6 @@ void PutDisplayEnv(u32 alp) {
 
 void aybabtu() {}
 
-void pc_set_levels(u32 lev_list) {
-  if (!Gfx::GetCurrentRenderer()) {
-    return;
-  }
-  std::vector<std::string> levels;
-  for (int i = 0; i < LEVEL_MAX; i++) {
-    u32 lev = *Ptr<u32>(lev_list + i * 4);
-    std::string ls = Ptr<String>(lev).c()->data();
-    if (ls != "none" && ls != "#f" && ls != "") {
-      levels.push_back(ls);
-    }
-  }
-
-  Gfx::GetCurrentRenderer()->set_levels(levels);
-}
-
-void pc_set_active_levels(u32 lev_list) {
-  if (!Gfx::GetCurrentRenderer()) {
-    return;
-  }
-  std::vector<std::string> levels;
-  for (int i = 0; i < LEVEL_MAX; i++) {
-    u32 lev = *Ptr<u32>(lev_list + i * 4);
-    std::string ls = Ptr<String>(lev).c()->data();
-    if (ls != "none" && ls != "#f" && ls != "") {
-      levels.push_back(ls);
-    }
-  }
-
-  Gfx::GetCurrentRenderer()->set_active_levels(levels);
-}
-
 //// PC Stuff
 void InitMachine_PCPort() {
   // PC Port added functions
@@ -396,11 +365,11 @@ void InitMachine_PCPort() {
       },
       make_string_from_c);
 
-  make_function_symbol_from_c("__pc-set-levels", (void*)pc_set_levels);
-  make_function_symbol_from_c("__pc-set-active-levels", (void*)pc_set_active_levels);
+  make_function_symbol_from_c("__pc-set-levels", (void*)kmachine_extras::pc_set_levels);
+  make_function_symbol_from_c("__pc-set-active-levels", (void*)kmachine_extras::pc_set_active_levels);
   make_function_symbol_from_c("__pc-get-tex-remap", (void*)lookup_jak3_texture_dest_offset);
   // make_function_symbol_from_c("pc-init-autosplitter-struct", (void*)init_autosplit_struct);
-  // make_function_symbol_from_c("pc-encode-utf8-string", (void*)encode_utf8_string);
+  make_function_symbol_from_c("pc-encode-utf8-string", (void*)kmachine_extras::encode_utf8_string);
 
   // discord rich presence
   // make_function_symbol_from_c("pc-discord-rpc-update", (void*)update_discord_rpc);
