+++ goal_src/jak1/pc/debug/pc-debug-common.gc
@@ -29,17 +29,19 @@
         (remain (the float ,remain))
         (bar-width (the int (/ (the float MEM_BAR_WIDTH) (-> *pc-settings* aspect-ratio-scale))))
         (bar-x (- MEM_BAR_RIGHT bar-width MEM_BAR_HORZ_PAD)) ;; x coord for left side of the bar list
-        (used-p (if (zero? total) 0.5 (/ (- total remain) total)))
+        (used-p (if (zero? total) 0.0 (/ (- total remain) total)))
         (used-x (the int (* used-p bar-width)))
         (used-y (+ MEM_BAR_Y (* ,idx MEM_BAR_HEIGHT)))
         )
       (draw-sprite2d-xy ,buf    bar-x         used-y used-x MEM_BAR_HEIGHT ,color)
       (draw-sprite2d-xy ,buf (+ bar-x used-x) used-y (- bar-width used-x) MEM_BAR_HEIGHT MEM_BAR_BG_COL)
+      (if (zero? total) (set! used-x (the int (* 0.5 bar-width))))
       (draw-string-xy ,name ,buf (- bar-x MEM_BAR_HORZ_PAD) used-y (font-color red) (font-flags shadow kerning right))
       (draw-string-xy (if (zero? total) "NO HEAP" (string-format "~,,2f%" (* used-p 100))) ,buf (+ bar-x used-x) used-y (font-color default) (font-flags shadow kerning middle))
       (draw-string-xy (string-format "~,,1fM" (/ total (* 1024 1024))) ,buf (+ bar-x bar-width MEM_BAR_HORZ_PAD) used-y (font-color red) (font-flags shadow kerning middle-vert))
       )
   )
+
 (defmacro draw-memory-bar-kheap (buf heap &key (name #f) &key idx &key color)
 	"draw a memory usage bar for a kheap"
   `(let ((heap ,heap))
