+++ goal_src/jak1/pc/pckernel-common.gc
@@ -387,7 +387,7 @@
   "return the debug font scale factor to be used."
   (declare (inline))
   (if (-> this debug-font-scale-auto?)
-      (/ (-> this debug-font-scale) (max 1.0 (/ (the float (get-current-game-height this)) PC_BASE_HEIGHT)))
+      (/ (-> this debug-font-scale) (fmax 1.0 (/ (the float (get-current-game-height this)) PC_BASE_HEIGHT)))
       (-> this debug-font-scale)))
