+++ goal_src/jak2/kernel/gstring.gc
@@ -456,6 +456,7 @@
     (cat-string<-string_to_charp arg0 arg1 s4-0)
     )
   )
+
 (defmacro is-whitespace-char? (c)
   ;; 32 = space
   ;; 9 = \t
