+++ goal_src/jak3/engine/debug/menu.gc
@@ -144,18 +144,18 @@
    (hilite-timer   int8)
    )
   (:methods
-    (new (symbol type string int (function object object)) _type_)
+    (new (symbol type string object (function object object)) _type_)
     )
   )
 
 
-(defmethod new debug-menu-item-function ((allocation symbol) (type-to-make type) (arg0 string) (arg1 int) (arg2 (function object object)))
+(defmethod new debug-menu-item-function ((allocation symbol) (type-to-make type) (arg0 string) (arg1 object) (arg2 (function object object)))
   (let ((v0-0 (object-new allocation type-to-make (the-as int (-> type-to-make size)))))
     (set! (-> v0-0 name) arg0)
     (set! (-> v0-0 parent) #f)
     (set! (-> v0-0 refresh-delay) 0)
     (set! (-> v0-0 refresh-ctr) (-> v0-0 refresh-delay))
-    (set! (-> v0-0 id) arg1)
+    (set! (-> v0-0 id) (the-as int arg1))
     (set! (-> v0-0 activate-func) arg2)
     (set! (-> v0-0 hilite-timer) 0)
     v0-0
@@ -167,7 +167,7 @@
    (is-on          symbol)
    )
   (:methods
-    (new (symbol type string int (function object debug-menu-msg object)) _type_)
+    (new (symbol type string object (function object debug-menu-msg object)) _type_)
     )
   )
 
@@ -175,7 +175,7 @@
 (defmethod new debug-menu-item-flag ((allocation symbol)
                                    (type-to-make type)
                                    (arg0 string)
-                                   (arg1 int)
+                                   (arg1 object)
                                    (arg2 (function object debug-menu-msg object))
                                    )
   (let ((v0-0 (object-new allocation type-to-make (the-as int (-> type-to-make size)))))
@@ -183,7 +183,7 @@
     (set! (-> v0-0 parent) #f)
     (set! (-> v0-0 refresh-delay) 23)
     (set! (-> v0-0 refresh-ctr) (-> v0-0 refresh-delay))
-    (set! (-> v0-0 id) arg1)
+    (set! (-> v0-0 id) (the-as int arg1))
     (set! (-> v0-0 activate-func) arg2)
     (set! (-> v0-0 is-on) #f)
     v0-0
@@ -585,7 +585,7 @@
                    'debug
                    'debug-menu-item-flag
                    (the-as string s4-1)
-                   (the-as int (car (cdr (cdr arg1))))
+                   (car (cdr (cdr arg1)))
                    (the-as (function object debug-menu-msg object) (debug-menu-func-decode (car (cdr (cdr (cdr arg1))))))
                    )
                  )
@@ -594,7 +594,7 @@
                    'debug
                    'debug-menu-item-function
                    (the-as string s4-1)
-                   (the-as int (car (cdr (cdr arg1))))
+                   (car (cdr (cdr arg1)))
                    (the-as (function object object) (debug-menu-func-decode (car (cdr (cdr (cdr arg1))))))
                    )
                  )
