+++ goal_src/jak3/engine/draw/drawable.gc
@@ -1957,6 +1957,12 @@
               (s1-7 *temp-string* s2-7 448 210 (font-color default) (font-flags shadow kerning))
               )
             )
+
+          (#when PC_PORT
+            (draw *pc-settings* s2-7)
+            (draw-memory *pc-settings* s2-7)
+            (print-debug-misc *pc-settings*)
+            )
           (display-file-info)
           )
         )
