+++ goal_src/jak3/engine/game/main.gc
@@ -1661,6 +1661,12 @@
       (with-profiler 'actors *profile-actors-color*
         (suspend)
         )
+      
+      ;; run the pc port hooks and code
+      (#when PC_PORT
+        (update *pc-settings*)
+        (if (and *display-sha* *debug-segment*)
+            (draw-build-revision)))
       )
     )
   (set! *dproc* #f)
