+++ goal_src/jak3/engine/level/level-h.gc
@@ -5,6 +5,11 @@
 ;; name in dgo: level-h
 ;; dgos: GAME
 
+;; max amount of levels in level heap
+(defconstant LEVEL_MAX 10)
+;; total amount of levels, including ones outside level heap (default-level)
+(defconstant LEVEL_TOTAL 11)
+
 (declare-type bsp-header basic)
 (declare-type drawable basic)
 (declare-type entity-links-array inline-array-class)
