+++ goal_src/jak3/kernel-defs.gc
@@ -184,7 +184,9 @@
 (define-extern pc-waiting-for-bind? (function symbol))
 (define-extern pc-set-waiting-for-bind! (function int symbol symbol int none))
 (define-extern pc-stop-waiting-for-bind! (function none))
+(define-extern pc-get-controller-index (function int int))
 (define-extern pc-set-controller! (function int int none))
+(define-extern pc-get-keyboard-enabled? (function symbol))
 (define-extern pc-set-keyboard-enabled! (function symbol none))
 (define-extern pc-set-mouse-options! (function symbol symbol symbol none))
 (define-extern pc-set-mouse-camera-sens! (function float float none))
@@ -197,16 +199,18 @@
 
 ;; Display Related Functions
 (define-extern pc-get-display-mode (function symbol))
-(define-extern pc-get-active-display-size (function (pointer int32) (pointer int32) none))
+(define-extern pc-get-active-display-size (function (pointer int64) (pointer int64) none))
 (define-extern pc-get-active-display-refresh-rate (function int))
 (define-extern pc-get-display-count (function int))
-(define-extern pc-get-display-name (function int string))
+(define-extern pc-get-display-name (function int string symbol))
 (define-extern pc-get-os (function symbol))
-(define-extern pc-get-window-size (function (pointer int32) (pointer int32) none))
+(define-extern pc-get-window-size (function (pointer int64) (pointer int64) none))
 (define-extern pc-get-window-scale (function (pointer float) (pointer float) none))
 (define-extern pc-set-window-size (function int int none))
 (define-extern pc-set-fullscreen-display (function int none))
 (define-extern pc-set-display-mode (function symbol none))
+(define-extern pc-get-num-resolutions (function int))
+(define-extern pc-get-resolution (function int (pointer int64) (pointer int64) none))
 
 (define-extern pc-set-frame-rate (function int none))
 (define-extern pc-set-vsync (function symbol none))
