+++ goal_src/jak3/kernel/gstring.gc
@@ -447,6 +447,22 @@
     )
   )
 
+(defmacro is-whitespace-char? (c)
+  ;; 32 = space
+  ;; 9 = \t
+  ;; 13 = \r
+  ;; 10 = \n
+  `(or (= ,c 32)
+       (= ,c 9)
+       (= ,c 13)
+       (= ,c 10)
+       )
+  )
+
+(defmacro not-whitespace-char? (c)
+  `(not (is-whitespace-char? ,c))
+  )
+
 (defun string-skip-whitespace ((arg0 (pointer uint8)))
   "Jump over whitespace chars."
   (while (and (nonzero? (-> arg0 0)) (or (= (-> arg0 0) 32) (= (-> arg0 0) 9) (= (-> arg0 0) 13) (= (-> arg0 0) 10)))
@@ -799,4 +815,24 @@
 
 (define *temp-string* (new 'global 'string 2048 (the-as string #f)))
 
+(#when PC_PORT (define *pc-encoded-temp-string* (new 'global 'string 2048 (the-as string #f))))
+
 (kmemclose)
+
+(defmacro string-format (&rest args)
+  "Formats into *temp-string* and returns it, for in-place string formating.
+   DO NOT USE *temp-string* WITH THIS MACRO! It is read as input AFTER all of the args evaluate."
+
+  `(begin
+     (format (clear *temp-string*) ,@args)
+     *temp-string*
+     )
+  )
+
+(defmacro temp-string-format (buf &rest args)
+  "Like [[string-format]], but takes a string as an argument."
+  `(begin
+     (format (clear ,buf) ,@args)
+     ,buf
+     )
+  )
